Collect Only Flag
This PR adds support for the `--collect-only` flag, which much like pytest
will run through and simply do test collection.  It also adds support for
timing the run as pytest does as well as adding e2e testing of the new
command.

It also removes a little unused code around watch mode and fixture collection
as well as renaming the `test` field on the `TestCase` struct to `name`
so it reads as `test.name` rather than `test.test`.

Closes #9