Introduce Module Organization (#17)
* Introduce Module Organization

It was getting a bit crowded having so much in our `lib.rs`, so this is
a first step at introducing some organization.  This will be compared to
a different approach in the next commit.

In addition this does a couple of other things:
* Adds a test for --help
* Removes the author part, as we now have more than one author
* Replaces our custom Rysult with use of anyhow

* Option Two for Module Organization

I slightly favor this but don't have a huge preference, so open to
input.