Simplify Directory CLI Args and Hopefully Improve Collection
This removes the dependency on the `glob` library and instead uses
a more simple plain directory-based method for determining where to
search for tests.  This should make it easier to supply the directories,
enhance compatibility with pytest, and maybe mean we actually pick up
all the tests we intend to?

There are undoubtedly some edge cases we haven't thought of here, but
this is probably a good start.

Also adds support for an easier to read test diffing experience.