In working on the release I saw an issue with accessing the vitualenv
Not sure why this hasn't come up before, but decided to extract the path
set up into its own thing so I could call it everywhere.  Also set up a little
info thing that we can use to start to print out debug information people might
want to use to file bug reports with - in this case I used it to ensure we saw
the right path.