Initial Release (#40)
* Initial Release

This is set to be our initial release of rytest.  I've done the following ahead of it:

- Bumped the version to 0.1.0.  Yay!
- Switched our release system to just use maturin's auto-generated [GHA](https://www.maturin.rs/distribution#github-actions).
- Added a document to record how to do a release, so we don't forget.

This should be everything that's needed, but we'll see.  :)

* Remove 3.13 support for now as PyO3 does not appear to support it, add missing benchmark images

* Update PyO3 to get 3.13 support back, make rytest the default cargo target

* In working on the release I saw an issue with accessing the vitualenv

Not sure why this hasn't come up before, but decided to extract the path
set up into its own thing so I could call it everywhere.  Also set up a little
info thing that we can use to start to print out debug information people might
want to use to file bug reports with - in this case I used it to ensure we saw
the right path.

* Add missing file

* Try some settings from https://github.com/messense/auditwheel-symbols/blob/master/.github/workflows/CI.yml

* Try switching to main

* I'm not confident this will fix it, but I think it might get us furhter.

Problem - we are missing the .so in the image.

This fixes that, but the reason for this is that we're trying to build a manylinux
compatible thingy, which means this might not work?  Dunno.

* Remove Linux (for now)