Add Rytest E2E
This little helper tool just allows us to quickly pull down and run tests
against popular repos.  For now I'll be using it to help track down issues
we have with collection and execution.  Longer term we'll probably want to
expand this and put it in CI to ensure there are no regressions, and we can
perhaps replace our custom bash scripts with some of this for driving our
benchmarks.